Feature/project endpoints

+++ BackendTascly/Repositories/ProjectsRepository.cs
@@ -19,37 +19,24 @@ namespace BackendTascly.Repositories
         public async Task<bool> DeleteProjectAsync(Guid projectId)
         {
             var project = await context.Projects.FindAsync(projectId);
+            if (project is null) return false;
+            
             context.Projects.Remove(project);
 
             var affected = await context.SaveChangesAsync();
             return affected > 0;
         }
 
-        public async Task<List<Project>> GetAllProjectsByOwnerId(Guid ownerId)
-        {
-            var projects = await context.Projects
-                //.Include(p => p.Members)
-                .Include(p => p.TaskStatuses)
-                .Include(p => p.TaskImportances)
-                .Include(p => p.Tasks)
-                //.Include(p => p.Roles)
-                .Include(p => p.Owner)
-                .Where(p => p.OwnerId == ownerId).ToListAsync();
-            return projects;
-        }
-
         public Task<Project?> GetProjectById(Guid projectId)
         {
-            var project = context.Projects
-                //.Include(p => p.Members)
-                .Include(p => p.TaskStatuses)
-                .Include(p => p.TaskImportances)
-                .Include(p => p.Tasks)
-                //.Include(p => p.Roles)
-                .Include(p => p.Owner)
-                .FirstOrDefaultAsync(p => p.Id == projectId);
+            var project = context.Projects.FirstOrDefaultAsync(p => p.Id == projectId);
 
             return project;
+        }
+
+        public async Task<List<Project>> GetProjectsByWorkspaceId(Guid workspaceId)
+        {
+            return await context.Projects.Where(p => p.WorkspaceId == workspaceId).ToListAsync();
         }
     }
 }
